Add MPAS-Analysis parameters for mapping in parallel
The new E3SM-Unified supports ESMF map-file generation in parallel
with system MPI on most machines.  (ESMF could not be built with system
MPI on Cori.)  On Anvil andChrysalis, MPAS-Analysis will not be able to
run unless ESMF calls are made with `srun` (which is not the
MPAS-Analysis default).

This merge updates the config options related to parallel map-file
generation to be consistent with the new MPAS-Analysis and E3SM-Unified.