Add campaign parameter
This merge adds a new "campaign" parameter to the "defaults" section
that is "custom" by default.  For other values (currently only
"water_cycle"), a config file is loaded with alternative default
parameters.  The campaign parameters take priority over the defaults
but the user parameters take priority over the campaign parameters.
Thus, a user could set the campaign to "water_cycle" but still
modify the set of analyses to run (e.g. for debugging).