Add campaign parameter (#138)
* Add campaign parameter

This merge adds a new "campaign" parameter to the "default" section
that is "none" by default.  For other values (currently only
"water_cycle"), a config file is loaded with alternative default
parameters.  The campaign parameters take priority over the defaults
but the user parameters take priority over the campaign parameters.
Thus, a user could set the campaign to "water_cycle" but still
modify the set of analyses to run (e.g. for debugging).

* Validate both user and campaign configs

The user config must be validated before we attempt to load the
campaign config in order to pick up the defaults.

* Rename default campaign from "custom" to "none"

* Tests and docs

* Format fix

Co-authored-by: Ryan Forsyth <forsyth2@llnl.gov>