Fix validation of campaign configs
This merge switches to first merging user and campaign config
files and only validating after the merge.  This slightly
complicates the logic for getting the campaign (if any) because
defaults have not yet been parsed at that point.  But it does
mean we only validate once (so only one validation success is
printed) and it means that defaults will not override the
campaign config options, as was happening before this merge.