Remove MPAS-Analysis ocean conservation from default config
We do not plan to run with the conservation check on by default,
so it doesn't make sense to have the analysis plots on by default.

Additionally, the v3 reference simulation is outputting conservation
at a monthly interval, but MPAS-Analysis assumes it is being
output at a daily interval, the new default.  Incorrect rolling
averages were leading to no data being plotted.