update the definition of the variable list to be flexible for ploting and viewer
This comment contains update in the default.ini interms of the
definition of "vars" in pcmdi diagnostic sections. I now explicitly
defined "clim_vars", "mova_vars", "movc_vars" and "enso_vars" to
explicitly pass the needed variables for corresponding diagnostics in
pcmdi. This will be later on used to process the viewer page

The corresponding changes were also made in pcmdi_diags.bash to accept
the new variables defined in .cfg files and default.ini